requirements.txt
Flask==2.0.1
google-cloud-logging==3.0.0
google-auth==2.0.2
